+++ sdk/ai/azure-ai-agents/azure/ai/agents/telemetry/_ai_agents_instrumentor.py
@@ -458,11 +458,7 @@ class _AIAgentsInstrumentorPreview:
                     t.type: t.bing_grounding,
                 }
             elif isinstance(t, RunStepOpenAPIToolCall):
-                tool_call = {
-                    "id": t.id,
-                    "type": t.type,
-                    'function': t.as_dict().get('function', {})
-                }
+                tool_call = {"id": t.id, "type": t.type, "function": t.as_dict().get("function", {})}
             elif isinstance(t, RunStepMcpToolCall):
                 tool_call = {
                     "id": t.id,
@@ -470,7 +466,7 @@ class _AIAgentsInstrumentorPreview:
                     "arguments": t.arguments,
                     "name": t.name,
                     "output": t.output,
-                    "server_label": t.server_label or ""
+                    "server_label": t.server_label or "",
                 }
             else:
                 # Works for Deep research
@@ -589,7 +585,7 @@ class _AIAgentsInstrumentorPreview:
     ) -> None:
         # TODO document new fields
 
-        event_body: dict[str, Any]= {}
+        event_body: dict[str, Any] = {}
         if _trace_agents_content:
             if isinstance(content, List):
                 for block in content:
@@ -1558,15 +1554,15 @@ class _AIAgentsInstrumentorPreview:
             if class_function_name.startswith("MessagesOperations.create"):
                 kwargs.setdefault("merge_span", True)
                 return self.trace_create_message(function, *args, **kwargs)
+            if class_function_name.startswith("RunsOperations.create_and_process"):
+                kwargs.setdefault("merge_span", True)
+                return self.trace_create_run(OperationName.PROCESS_THREAD_RUN, function, *args, **kwargs)
             if class_function_name.startswith("RunsOperations.create"):
                 kwargs.setdefault("merge_span", True)
                 return self.trace_create_run(OperationName.START_THREAD_RUN, function, *args, **kwargs)
             if class_function_name.startswith("RunsOperations.get"):
                 kwargs.setdefault("merge_span", True)
                 return self.trace_get_run(OperationName.GET_THREAD_RUN, function, *args, **kwargs)
-            if class_function_name.startswith("RunsOperations.create_and_process"):
-                kwargs.setdefault("merge_span", True)
-                return self.trace_create_run(OperationName.PROCESS_THREAD_RUN, function, *args, **kwargs)
             if class_function_name.startswith("AgentsClient.create_thread_and_run"):
                 kwargs.setdefault("merge_span", True)
                 return self.trace_create_run(OperationName.PROCESS_THREAD_RUN, function, *args, **kwargs)
@@ -1634,15 +1630,15 @@ class _AIAgentsInstrumentorPreview:
             if class_function_name.startswith("MessagesOperations.create"):
                 kwargs.setdefault("merge_span", True)
                 return await self.trace_create_message_async(function, *args, **kwargs)
+            if class_function_name.startswith("RunsOperations.create_and_process"):
+                kwargs.setdefault("merge_span", True)
+                return await self.trace_create_run_async(OperationName.PROCESS_THREAD_RUN, function, *args, **kwargs)
             if class_function_name.startswith("RunsOperations.create"):
                 kwargs.setdefault("merge_span", True)
                 return await self.trace_create_run_async(OperationName.START_THREAD_RUN, function, *args, **kwargs)
             if class_function_name.startswith("RunsOperations.get"):
                 kwargs.setdefault("merge_span", True)
                 return await self.trace_get_run_async(OperationName.GET_THREAD_RUN, function, *args, **kwargs)
-            if class_function_name.startswith("RunsOperations.create_and_process"):
-                kwargs.setdefault("merge_span", True)
-                return await self.trace_create_run_async(OperationName.PROCESS_THREAD_RUN, function, *args, **kwargs)
             if class_function_name.startswith("AgentsClient.create_thread_and_run"):
                 kwargs.setdefault("merge_span", True)
                 return await self.trace_create_run_async(OperationName.PROCESS_THREAD_RUN, function, *args, **kwargs)
@@ -2101,7 +2097,9 @@ class _AgentEventHandlerTraceWrapper(AgentEventHandler):
         # See work item 4636616 and 4636299 for details.
         # When the work item is resolved, change this code back to:
         # if message.status in {MessageStatus.COMPLETED, MessageStatus.INCOMPLETE}
-        if message.status in {MessageStatus.COMPLETED, MessageStatus.INCOMPLETE} or (message.status == MessageStatus.IN_PROGRESS and message.content):
+        if message.status in {MessageStatus.COMPLETED, MessageStatus.INCOMPLETE} or (
+            message.status == MessageStatus.IN_PROGRESS and message.content
+        ):
             self.last_message = message
 
         return retval  # type: ignore
@@ -2242,7 +2240,9 @@ class _AsyncAgentEventHandlerTraceWrapper(AsyncAgentEventHandler):
         # See work item 4636616 and 4636299 for details.
         # When the work item is resolved, change this code back to:
         # if message.status in {MessageStatus.COMPLETED, MessageStatus.INCOMPLETE}
-        if message.status in {MessageStatus.COMPLETED, MessageStatus.INCOMPLETE} or (message.status == MessageStatus.IN_PROGRESS and message.content):
+        if message.status in {MessageStatus.COMPLETED, MessageStatus.INCOMPLETE} or (
+            message.status == MessageStatus.IN_PROGRESS and message.content
+        ):
             self.last_message = message
 
         return retval  # type: ignore