+++ sdk/identity/azure-identity/azure/identity/_internal/__init__.py
@@ -9,6 +9,7 @@ from .aadclient_certificate import AadClientCertificate
 from .decorators import wrap_exceptions
 from .interactive import InteractiveCredential
 from .utils import (
+    encode_base64,
     get_default_authority,
     normalize_authority,
     process_credential_exclusions,
@@ -46,6 +47,7 @@ __all__ = [
     "AadClientBase",
     "AuthCodeRedirectServer",
     "AadClientCertificate",
+    "encode_base64",
     "get_default_authority",
     "InteractiveCredential",
     "normalize_authority",