+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 1.0.0b42 (Unreleased)
+## 1.0.0b43 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0b42 (2025-09-17)
 
 ### Features Added
 - Customer Facing Statsbeat: Added remaining drop codes to base
@@ -15,10 +25,12 @@
   ([#42573](https://github.com/Azure/azure-sdk-for-python/pull/42573))
 - Customer Facing SDKStats: Exception categorization as per [Spec] - https://github.com/aep-health-and-standards/Telemetry-Collection-Spec/blob/main/ApplicationInsights/sdkstats/customer_facing_sdk_stats.md
   ([#42695](https://github.com/Azure/azure-sdk-for-python/pull/42695))
-
-### Breaking Changes
-
-### Bugs Fixed
+- Customer Facing SDKStats: Added logic for race conditions and updated the implementation to use a global instance for customer SDKStats metrics
+  ([#42655](https://github.com/Azure/azure-sdk-for-python/pull/42655))
+- Customer Facing SDKStats: Added telemetry_success field to dropped items as per [Spec] - https://github.com/aep-health-and-standards/Telemetry-Collection-Spec/pull/606
+  ([#42846](https://github.com/Azure/azure-sdk-for-python/pull/42846))
+- Customer Facing SDKStats: Refactor to use `Manager` and `Singleton` pattern
+  ([#42969](https://github.com/Azure/azure-sdk-for-python/pull/42969))
 
 ### Other Changes
 
@@ -28,6 +40,8 @@
   ([#42508] https://github.com/Azure/azure-sdk-for-python/pull/42508)
 - Refactoring of statsbeat to use `StatsbeatManager`
   ([#42716] https://github.com/Azure/azure-sdk-for-python/pull/42716)
+- Support sdk stats dynamic change on OneSettings config change
+  ([#42877] https://github.com/Azure/azure-sdk-for-python/pull/42877)
 
 ## 1.0.0b41 (2025-07-31)